sell/src/main/java/cn/yzy/service/impl/ProductServiceImpl.java
package cn.yzy.service.impl;

import cn.yzy.dataobject.ProductInfo;
import cn.yzy.dto.CartDTO;
import cn.yzy.enums.ProductStatusEnum;
import cn.yzy.enums.ResultEnum;
import cn.yzy.exception.SellException;
import cn.yzy.repostitory.ProductinfoRepository;
import cn.yzy.service.ProductService;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.data.domain.Page;
import org.springframework.data.domain.Pageable;
import org.springframework.stereotype.Service;

import javax.transaction.Transactional;
import java.util.List;
import java.util.Optional;

@Service
public class ProductServiceImpl implements ProductService {
    @Autowired
    private ProductinfoRepository repository;
    @Override
    public ProductInfo findOne(String productId) {
//        return repository.findById(productId).get();
//        return repository.getOne(productId);
        Optional<ProductInfo> optional = repository.findById(productId);
        return optional.isPresent() ? optional.get() : null;
    }




    @Override
    public List<ProductInfo> findUpAll() {
        return repository.findByProductStatus(ProductStatusEnum.UP.getCode());
    }

    @Override
    public Page<ProductInfo> findAll(Pageable pageable) {
        return repository.findAll(pageable);
    }

    @Override
    public ProductInfo save(ProductInfo productInfo) {
        return repository.save(productInfo);
    }

    @Override
    @Transactional
    public void increaseStock(List<CartDTO> cartDTOList) {
        for (CartDTO cartDTO : cartDTOList) {
            ProductInfo productInfo = repository.findById(cartDTO.getProductId()).get();
            if(productInfo == null){
                throw new SellException(ResultEnum.PRODUCT_NOT_EXIT);
            }
            Integer result = productInfo.getProductStock() + cartDTO.getProductQuantity();
            productInfo.setProductStock(result);
            repository.save(productInfo);
        }

    }
    @Override
    @Transactional
    public void decreaseStock(List<CartDTO> cartDTOList) {
        for( CartDTO cartDTO:cartDTOList){
            ProductInfo productInfo = repository.findById(cartDTO.getProductId()).get();
            if(productInfo == null){
                throw new SellException(ResultEnum.PRODUCT_NOT_EXIT);
            }
            Integer result = productInfo.getProductStock() - cartDTO.getProductQuantity();
            if(result < 0){
                throw new SellException(ResultEnum.PRODUCT_STOCK_ERROR);
            }
            productInfo.setProductStock(result);
            repository.save(productInfo);
        }
    }
}
